+++ lib/events/auditlog.go
@@ -1045,7 +1045,7 @@ func (l *AuditLog) SearchSessionEvents(fromUTC, toUTC time.Time, limit int, orde
 // StreamSessionEvents streams all events from a given session recording. An error is returned on the first
 // channel if one is encountered. Otherwise it is simply closed when the stream ends.
 // The event channel is not closed on error to prevent race conditions in downstream select statements.
-func (l *AuditLog) StreamSessionEvents(ctx context.Context, sessionID session.ID, startIndex int) (chan apievents.AuditEvent, chan error) {
+func (l *AuditLog) StreamSessionEvents(ctx context.Context, sessionID session.ID, startIndex int64) (chan apievents.AuditEvent, chan error) {
 	l.log.Debugf("StreamSessionEvents(%v)", sessionID)
 	e := make(chan error, 1)
 	c := make(chan apievents.AuditEvent)
@@ -1317,7 +1317,7 @@ func (a *closedLogger) Close() error {
 	return trace.NotImplemented(loggerClosedMessage)
 }
 
-func (a *closedLogger) StreamSessionEvents(_ctx context.Context, sessionID session.ID, startIndex int) (chan apievents.AuditEvent, chan error) {
+func (a *closedLogger) StreamSessionEvents(_ctx context.Context, sessionID session.ID, startIndex int64) (chan apievents.AuditEvent, chan error) {
 	c, e := make(chan apievents.AuditEvent), make(chan error, 1)
 	e <- trace.NotImplemented(loggerClosedMessage)
